make PROJ library optional
If the proj library cannot be found by cmake, do not offer
the option to use arbitrary projections.

Also changes reprojection interface to return managed pointers
for the projection.

Fixes #1065.
