+++ src/reprojection-generic-none.cpp
@@ -1,0 +1,8 @@
+#include <stdexcept>
+
+#include "reprojection.hpp"
+
+std::shared_ptr<reprojection> reprojection::make_generic_projection(int srs)
+{
+    throw std::runtime_error("No generic projection library available.");
+}
